requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# HTTP & API
requests==2.31.0
aiofiles==23.2.1
python-multipart==0.0.6

# AI & ML
openai==1.3.7

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Templates & UI
jinja2==3.1.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
bcrypt==4.0.1

# Scheduling & Background Tasks
schedule==1.2.0
APScheduler==3.10.4

# Data Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0

# Data Processing  
pandas==2.1.1
numpy==1.24.3
scipy==1.11.3

# Utilities
python-multipart==0.0.6
python-dateutil==2.8.2
pytz==2023.3
uuid==1.30